MAINT: Remove bandit API for 1.0.0 release
This is a breaking change for anyone who hasn't migrated to the Agent API, but we've shown three major versions of deprecation warnings about it
